requirements.txt
extra_streamlit_components==0.1.53
st_annotated_text==3.0.0
streamlit==1.8.1
cryptography==36.0.1
streamlit-lottie==0.0.3
pandas
streamlit-aggrid==0.2.3.post2
geopy==2.2.0
requests
folium==0.12.1.post1
streamlit-folium==0.6.7
hydralit-components==1.0.9
google-cloud-firestore==2.4.0
cryptography==36.0.1